chore: improve windows support for scripts

+++ init-dev.sh
@@ -4,19 +4,36 @@
 
 set -e
 
+# import command functions
+. ./src/shell/_command.sh
+
+# import node functions
+. ./src/shell/_node.sh
+
 #######################################
 # Initializes node version manager (nvm).
 # Arguments:
 #   None
 #######################################
-initNvm() {
-    set +e
-    . "$HOME/.nvm/nvm.sh" # Load nvm
-    set -e
+init_nvm() {
+  if [ ! -s "$HOME/.nvm/nvm.sh" ]; then
+    echo '' 1>&2
+    echo 'ERROR: Missing required nvm program.' 1>&2
+    exit 1
+  fi
 
-    nvm_version=$(nvm --version)
+  set +e
+  . "$HOME/.nvm/nvm.sh" # Load nvm
+  set -e
 
-    echo "Detected nvm version: $nvm_version"
+  if ! require nvm; then
+    echo 'The CLI for nvm commands could not be found and must be installed.' 1>&2
+    exit 1
+  fi
+
+  nvm_version=$(nvm --version)
+
+  echo "Detected nvm version: $nvm_version"
 }
 
 #######################################
@@ -24,19 +41,60 @@ initNvm() {
 # Arguments:
 #   None
 #######################################
-initNode() {
-    node_version=$(cat ./.nvmrc)
-    node_version="${node_version%%*( )}"
+init_node() {
+
+  current_node_version=$(get_node_version)
+  target_node_version=$(get_target_node_version)
+
+  if [ "$current_node_version" = "$target_node_version" ]; then
+    echo "The \`node --version\` already matches the project version \"$target_node_version\"."
+  else
+    echo ''
+    echo "A different version of NodeJS was detected: $current_node_version."
+    echo "Installing NodeJS $target_node_version..."
 
     set +e
-    echo ''
-    echo "Installing NodeJS $node_version..."
-    nvm install $node_version
-    echo "Successfully installed NodeJS $node_version."
-    echo "Switching to NodeJS $node_version..."
-    nvm use $node_version
-    echo "Successfully switched to NodeJS $node_version."
+    nvm_install_output=$(nvm install $target_node_version 2>&1)
+    nvm_install_exit_code=$?
     set -e
+
+    if bad_exit "$nvm_install_exit_code" || contains "$nvm_install_output" "Failure" || contains "$nvm_install_output" "failed" || contains "$nvm_install_output" "not supported"; then
+      echo "$nvm_install_output" 1>&2
+      echo '' 1>&2
+      echo 'NodeJS installation failed.' 1>&2
+      echo 'Subsequent commands may fail.' 1>&2
+      echo "Try: \"nvm install $target_node_version\" and then re-run this command." 1>&2
+      echo '*Windows: this may require "Run as administrator".' 1>&2
+      # IMPORTANT: continue, do not stop
+    else
+      echo "Successfully installed NodeJS $target_node_version."
+    fi
+
+    echo ''
+    echo "Switching to NodeJS $target_node_version..."
+
+    set +e
+    nvm_use_output=$(nvm use $target_node_version 2>&1)
+    nvm_use_exit_code=$?
+    set -e
+
+    if bad_exit "$nvm_use_exit_code" || contains "$nvm_use_output" "not installed" || contains "$nvm_use_output" "denied"; then
+      echo "$nvm_use_output" 1>&2
+      echo '' 1>&2
+      echo 'Switching NodeJS version failed.' 1>&2
+      echo 'Subsequent commands may fail.' 1>&2
+      echo "Try: \"nvm use $target_node_version\" and then re-run this command." 1>&2
+      echo '*Windows: this may require "Run as administrator".' 1>&2
+      # IMPORTANT: continue, do not stop
+    else
+      echo "Successfully switched to NodeJS $target_node_version."
+    fi
+  fi
+
+  if ! command -v node >/dev/null 2>&1; then
+    echo 'The CLI for node commands could not be found and must be installed.' 1>&2
+    exit 1
+  fi
 }
 
 #######################################
@@ -44,14 +102,20 @@ initNode() {
 # Arguments:
 #   None
 #######################################
-initNpm() {
-    echo ''
-    echo "Installing latest npm..."
-    npm install --location=global --no-fund npm@latest
-    echo "Successfully installed latest npm."
+init_npm() {
 
-    npm config set fund false --global
-    echo 'Disabled npm funding messages.'
+  if ! require npm; then
+    echo 'The CLI for npm commands could not be found and must be installed.' 1>&2
+    exit 1
+  fi
+
+  echo ''
+  echo "Installing latest npm..."
+  npm install --location=global --no-fund npm@latest
+  echo "Successfully installed latest npm."
+
+  npm config set fund false --global
+  echo 'Disabled npm funding messages.'
 }
 
 #######################################
@@ -59,19 +123,38 @@ initNpm() {
 # Arguments:
 #   None
 #######################################
-initPnpm() {
-    pnpm_version=$(cat ./.pnpmrc)
-    pnpm_version="${pnpm_version%%*( )}"
+init_pnpm() {
 
-    echo ''
-    echo "Removing any previous version of pnpm..."
-    npm uninstall --location=global pnpm || true
-    echo "Successfully removed any previous version of pnpm."
+  # installing pnpm requires npm
+  if ! require npm; then
+    echo 'The CLI for npm commands could not be found and must be installed.' 1>&2
+    exit 1
+  fi
 
+  current_pnpm_version=$(get_pnpm_version)
+  target_pnpm_version=$(get_target_pnpm_version)
+
+  pnpm_install_required=1
+
+  if require pnpm; then
+    if [ "$current_pnpm_version" = "$target_pnpm_version" ]; then
+      echo "The \`pnpm --version\` already matches the project version \"$target_pnpm_version\"."
+      pnpm_install_required=0
+    else
+      echo ''
+      echo "Current \`pnpm --version\` does not match the project version \"$target_pnpm_version\"."
+      echo "Removing the previous version of pnpm..."
+      npm uninstall --location=global pnpm || true
+      echo "Successfully removed the previous version of pnpm."
+    fi
+  fi
+
+  if [ $pnpm_install_required = 1 ]; then
     echo ''
-    echo "Installing pnpm $pnpm_version..."
-    npm install --location=global pnpm@$pnpm_version
-    echo "Successfully installed pnpm $pnpm_version."
+    echo "Installing pnpm $target_pnpm_version..."
+    npm install --location=global pnpm@$target_pnpm_version
+    echo "Successfully installed pnpm $target_pnpm_version."
+  fi
 }
 
 #######################################
@@ -80,10 +163,10 @@ initPnpm() {
 #   None
 #######################################
 main() {
-    initNvm
-    initNode
-    initNpm
-    initPnpm
+  init_nvm
+  init_node
+  init_npm
+  init_pnpm
 }
 
 main
